+++ source/Public/Block/Add-NotionBlockToPage.ps1
@@ -82,8 +82,11 @@ function Add-NotionBlockToPage
     {
         Write-Debug "Adding block(s) to page $($Page.id)"
         $pageID ??= $Page.id
-        $blocks = $block | ConvertTo-Json -Depth 10 -EnumsAsStrings
-        $response =  Invoke-NotionApiCall -Uri "/blocks/$($Page.id)/children" -Method PATCH -body $blocks
+        $body = @{
+            children = @($Block)
+        }
+        $body = $body | Remove-NullValuesFromObject
+        $response =  Invoke-NotionApiCall -Uri "/blocks/$($Page.id)/children" -Method PATCH -body $body
         return [notion_block[]]@($response.foreach({ [notion_block]::ConvertFromObject($_) }))
     }
 }