+++ source/Public/Block/Heading/Add-NotionHeaderToBlock.ps1
@@ -37,6 +37,7 @@ function Add-NotionHeaderToBlock
     This example creates a new header object and then adds it to the specified parent block using the InputObject parameter.
     #>
     [CmdletBinding()]
+    [Alias("Add-NotionHeaderToPage")]
     param (
         [Parameter(Mandatory = $true, ParameterSetName = "Params", HelpMessage = "Text of the Header")]
         [string] $Text,